+++ include/DecisionTree.hpp
@@ -14,33 +14,34 @@ class DecisionTreeNode;
 class DecisionTree
 {
 public:
+    //--------------- Constructors and Destructors ----------------//
     DecisionTree(std::map<std::string, std::string> kwargs); // constructor
     ~DecisionTree(); // destructor
 
-    // Functions
+    //--------------- Functions ----------------//
     void getTrainingData();
     void calculateFirstOrderProbabilities();
     void showTrainingData() const;
 
 
-    // Classify
+    //--------------- Classify ----------------//
     std::map<std::string, std::string> classify(void* root_node, const std::vector<std::string>& features_and_values);
 
 
-    // Construct tree
+    //--------------- Construct Tree ----------------//
     DecisionTreeNode* constructDecisionTreeClassifier();
 
 
-    // Entropy calculators
+    //--------------- Entropy Calculators ----------------//
     double classEntropyOnPriors();
 
 
-    // Probability calculators
+    //--------------- Probability Calculators ----------------//
     double probabilityOfFeatureValue(const std::string& feature, const std::string& value);
     double probabilityOfFeatureValue(const std::string& feature, double sampling_point);
 
 
-    // Class based utilities
+    //--------------- Class Based Utilities ----------------//
 
 
     int _nodesCreated;
