--- src/Classify.cpp
@@ -1,5 +1,0 @@
-#include "DecisionTree.hpp"
-
-std::map<std::string, std::string> DecisionTree::classify(void* root_node, const std::vector<std::string>& features_and_values) {
-    return {};
-}
