--- src/ConstructTree.cpp
@@ -1,5 +1,0 @@
-#include "DecisionTree.hpp"
-
-DecisionTreeNode* DecisionTree::constructDecisionTreeClassifier() {
-    return nullptr;
-}
