--- src/EntropyCalculators.cpp
@@ -1,5 +1,0 @@
-#include "DecisionTree.hpp"
-
-double DecisionTree::classEntropyOnPriors() {
-    return 0.0;
-}
