--- src/ProbabilityCalculators.cpp
@@ -1,9 +1,0 @@
-#include "DecisionTree.hpp"
-
-double DecisionTree::probabilityOfFeatureValue(const std::string& feature, const std::string& value) {
-    return 1.0;
-}
-
-double DecisionTree::probabilityOfFeatureValue(const std::string& feature, double sampling_point) {
-    return 1.0;
-}
